+++ src/Tree/NativeMutable.php
@@ -56,7 +56,7 @@ class NativeMutable implements
      * Value based construct
      */
     public function __construct(
-        iterable $items = null,
+        ?iterable $items = null,
         mixed $value = null
     ) {
         if (!is_iterable($value)) {
@@ -678,7 +678,7 @@ class NativeMutable implements
      */
     public function toDelimitedSet(
         bool $urlEncode = false,
-        string $prefix = null
+        ?string $prefix = null
     ): array {
         $output = [];
 
@@ -910,6 +910,7 @@ class NativeMutable implements
      */
     public function getChildren(): array
     {
+        // @phpstan-ignore-next-line
         return $this->items;
     }
 
@@ -917,10 +918,11 @@ class NativeMutable implements
     /**
      * Iterator interface
      *
-     * @return Iterator<int|string, static<TValue>>
+     * @return Iterator<int|string,static<TValue>>
      */
     public function getIterator(): Iterator
     {
+        // @phpstan-ignore-next-line
         return new ArrayIterator($this->items);
     }
 
@@ -968,13 +970,14 @@ class NativeMutable implements
     /**
      * Copy and reinitialise new object
      *
-     * @param iterable<int|string, TValue|iterable<mixed>> $newItems
+     * @param iterable<int|string,TValue|iterable<mixed>> $newItems
      * @param TValue|null $value
      */
     protected static function propagate(
         ?iterable $newItems = [],
         mixed $value = null
     ): static {
+        // @phpstan-ignore-next-line
         return new static($newItems, $value);
     }
 }